--- a/exmath/round.go
+++ b/exmath/round.go
@@ -11,5 +11,5 @@ import (
 // 返回将 val 根据指定精度 precision（十进制小数点后数字的数目）进行四舍五入的结果。precision 也可以是负数或零。
 func Round(val float64, precision int) float64 {
 	p := math.Pow10(precision)
-	return math.Floor((val+(0.5/p))*p) / p
+	return math.Floor(val*p+0.5) / p
 }