--- a/exmath/round_test.go
+++ b/exmath/round_test.go
@@ -65,3 +65,4 @@ func TestRound(t *testing.T) {
 		}
 	}
 }
+